+++ manuchat/src/components/UserListItem/UserListItem.js
@@ -1,5 +1,5 @@
-import { AddUser } from '../../services/action/addUser/AddUser'
-import { AddVisibility } from '../../services/action/addVisibility/AddVisibility';
+import { AddIdContact } from '../../services/actions/addIdContact/AddIdContact';
+import { AddVisibility } from '../../services/actions/addVisibility/AddVisibility';
 import store from '../../services/store/store';
 
 const UserListItem = (props) => {
@@ -8,12 +8,8 @@ const UserListItem = (props) => {
     const handleContactUser = (e) => {
         const contact =JSON.parse(e.target.parentNode.dataset.user);
         const id = contact._id;
-        const name = contact.name;
-        const surname = contact.surname;
-        const username = contact.username;
-        const email = contact.email;
         if (contact) {
-            store.dispatch(AddUser(id, name, surname, username, email));
+            store.dispatch(AddIdContact(id));
             store.dispatch(AddVisibility(true));
             
         };
