+++ manuchat/src/pages/login/Login.scss
@@ -1,4 +1,4 @@
-.container{
+.container__login{
     margin: 30rem auto; 
     max-width: 45rem;
     width: calc(100% - 40px);
