--- manuchat/src/services/action/addUser/AddUser.js
@@ -1,13 +1,0 @@
-export const AddUser = (id, name, surname, username, email) => {
-    return({
-        type: "ADD_USER",
-        payload:{
-            "id": id,
-            "name": name,
-            "surname": surname,
-            "username": username,
-            "email": email
-        }
-    })
-};
-
