+++ manuchat/src/services/actions/addIdContact/AddIdContact.js
@@ -1,0 +1,9 @@
+export const AddIdContact = (id) => {
+    return({
+        type: "ADD_IDCONTACT",
+        payload:{
+            "id": id
+        }
+    })
+};
+
